extract shared brief render contracts
